Added ticket, refund, changed price from float to int(bacause stripe uses int)

+++ PraksaProjektBackend/Models/Event.cs
@@ -16,7 +16,7 @@ namespace PraksaProjektBackend.Models
 
         public string? OrganizersName { get; set; }
 
-        public float Profit { get; set; } = 0;
+        public int Profit { get; set; } = 0;
 
         public int? EventTypeId { get; set; }
